refactor: refactor the query folder into simple_executor
