+++ rust/lance-graph/src/lib.rs
@@ -45,6 +45,7 @@ pub mod parser;
 pub mod query;
 pub mod query_processor;
 pub mod semantic;
+pub mod simple_executor;
 pub mod source_catalog;
 pub mod sql_converter;
